backend/requirements.txt
fastapi
uvicorn
redis
requests
pydantic
python-dotenv